refactor: Move dialogs to separate package
This commit moves the `BankCardDialog`, `RiskCardDialog`, and `StartBonusDialog` classes into a new `dialog` package within the `game` module.

This improves code organization and separation of concerns.
