+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/game/dialog/BankCardDialog.kt
@@ -1,4 +1,4 @@
-package at.aau.serg.websocketbrokerdemo.game
+package at.aau.serg.websocketbrokerdemo.game.dialog
 
 import android.app.Dialog
 import android.content.Context
@@ -7,7 +7,7 @@ import android.widget.Button
 import android.widget.TextView
 import at.aau.serg.websocketbrokerdemo.R
 
-class RiskCardDialog(
+class BankCardDialog(
     context: Context,
     private val title: String,
     private val description: String
@@ -15,12 +15,12 @@ class RiskCardDialog(
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.dialog_risk_card)
+        setContentView(R.layout.dialog_bank_card)
 
-        findViewById<TextView>(R.id.riskCardTitle).text = title
-        findViewById<TextView>(R.id.riskCardText).text = description
+        findViewById<TextView>(R.id.bankCardTitle).text = title
+        findViewById<TextView>(R.id.bankCardText).text = description
 
-        findViewById<Button>(R.id.okRiskCardBtn).setOnClickListener {
+        findViewById<Button>(R.id.okBankCardBtn).setOnClickListener {
             dismiss()
         }
     }
